+++ Imperium/src/Core/ObjectManager.cs
@@ -38,6 +38,10 @@ internal class ObjectManager : ImpLifecycleObject
         RefreshLevelObstacles();
 
         LogObjects();
+        
+        // Reload objects that are hidden on the moon but visible in space
+        ImpSettings.Rendering.SpaceSun.Refresh();
+        ImpSettings.Rendering.StarsOverlay.Refresh();
     }
 
     protected override void OnPlayersUpdate(int playersConnected) => FetchPlayers();