+++ Imperium/src/MonoBehaviours/ImpUI/Common/ImpToggle.cs
@@ -1,5 +1,6 @@
 #region
 
+using System;
 using System.Linq;
 using Imperium.Core;
 using Imperium.Util;
@@ -46,12 +47,8 @@ public abstract class ImpToggle
 
         toggle.isOn = valueBinding.Value;
         toggle.onValueChanged.AddListener(valueBinding.Set);
-        toggle.onValueChanged.AddListener(_ =>
-        {
-            Imperium.Log.LogInfo($"[AAA] PLAYING SOUND FROM TOGGLE22: {path}");
-            GameManager.PlayClip(ImpAssets.GrassClick);
-        });
-        valueBinding.onUpdate += value => toggle.isOn = value;
+        toggle.onValueChanged.AddListener(_ => GameManager.PlayClip(ImpAssets.GrassClick));
+        valueBinding.syncOnUpdate += value => toggle.isOn = value;
 
         if (interactableBindings.Length > 0)
         {